refactor: update data structure for appeal
