+++ src/models/Appeal.ts
@@ -1,0 +1,9 @@
+import { PenaltyIdentifier } from "./PenaltyIdentifier";
+import { Reasons } from "./Reasons";
+
+export interface Appeal {
+  
+  penaltyIdentifier: PenaltyIdentifier;
+  reasons: Reasons;
+
+}
