+++ src/models/OtherReasons.ts
@@ -1,0 +1,6 @@
+export interface OtherReasons {
+  
+  title: string;
+  description: string;
+  
+}
