+++ src/models/Reasons.ts
@@ -1,0 +1,7 @@
+import { OtherReason } from './OtherReason';
+
+export interface Reasons {
+
+  other: OtherReason;
+  
+}
