+++ test/controllers/PenaltyDetailsController.test.ts
@@ -7,7 +7,7 @@ import '../../src/controllers/PenaltyDetailsController';
 import { RedisService } from '../../src/services/RedisService';
 import { MOVED_TEMPORARILY, OK, UNPROCESSABLE_ENTITY } from 'http-status-codes';
 import { expect } from 'chai';
-import { PenaltyReferenceDetails } from '../../src/models/PenaltyReferenceDetails';
+import { PenaltyIdentifier } from '../../src/models/PenaltyIdentifier';
 import { PENALTY_DETAILS_PAGE_URI, OTHER_REASON_DISCLAIMER_PAGE_URI } from '../../src/utils/Paths'
 
 const pageHeading = 'What are the penalty details?';
@@ -30,14 +30,14 @@ describe('PenaltyDetailsController', () => {
         });
 
         it('should return 200 when trying to access page with a session', async () => {
-            const penaltyDetails: Record<string, any> = {
+            const penaltyIdentifier: Record<string, any> = {
                 penaltyReference: 'A12345678',
                 companyNumber: 'SC123123'
             };
 
             const app = createApplication(container => {
                 container.bind(RedisService).toConstantValue(createSubstituteOf<RedisService>(service => {
-                    service.getObject('1').returns(Promise.resolve(penaltyDetails))
+                    service.getObject('1').returns(Promise.resolve(penaltyIdentifier))
                 }));
             });
 
@@ -46,8 +46,8 @@ describe('PenaltyDetailsController', () => {
                 .expect(response => {
                     expect(response.status).to.be.equal(OK);
                     expect(response.text).to.contain(pageHeading)
-                        .and.to.contain(penaltyDetails.companyNumber)
-                        .and.to.contain(penaltyDetails.penaltyReference)
+                        .and.to.contain(penaltyIdentifier.companyNumber)
+                        .and.to.contain(penaltyIdentifier.penaltyReference)
                         .and.not.contain(errorSummaryHeading)
                 })
         });
@@ -55,19 +55,19 @@ describe('PenaltyDetailsController', () => {
 
     describe('POST request', () => {
         it('should return 302 and redirect to disclaimer page when posting valid penalty details', async () => {
-            const penaltyDetails: PenaltyReferenceDetails = {
+            const penaltyIdentifier: PenaltyIdentifier = {
                 penaltyReference: 'A12345678',
                 companyNumber: 'SC123123'
             };
 
             const app = createApplication(container => {
                 container.bind(RedisService).toConstantValue(createSubstituteOf<RedisService>(service => {
-                    service.setObject('1', penaltyDetails).returns(Promise.resolve())
+                    service.setObject('1', penaltyIdentifier).returns(Promise.resolve())
                 }));
             });
 
             await request(app).post(PENALTY_DETAILS_PAGE_URI)
-                .send(penaltyDetails)
+                .send(penaltyIdentifier)
                 .expect(response => {
                     expect(response.status).to.be.equal(MOVED_TEMPORARILY);
                     expect(response.get('Location')).to.be.equal(OTHER_REASON_DISCLAIMER_PAGE_URI);
@@ -76,13 +76,13 @@ describe('PenaltyDetailsController', () => {
         });
 
         it('should return 400 when posting empty penalty reference', async () => {
-            const penaltyDetails: PenaltyReferenceDetails = {
+            const penaltyIdentifier: PenaltyIdentifier = {
                 penaltyReference: '',
                 companyNumber: 'SC123123'
             };
 
             await request(app).post(PENALTY_DETAILS_PAGE_URI)
-                .send(penaltyDetails)
+                .send(penaltyIdentifier)
                 .expect(response => {
                     expect(response.status).to.be.equal(UNPROCESSABLE_ENTITY);
                     expect(response.text).to.contain(pageHeading)
@@ -91,13 +91,13 @@ describe('PenaltyDetailsController', () => {
         });
 
         it('should return 400 when posting invalid penalty reference', async () => {
-            const penaltyDetails: PenaltyReferenceDetails = {
+            const penaltyIdentifier: PenaltyIdentifier = {
                 penaltyReference: '12345678',
                 companyNumber: 'SC123123'
             };
 
             await request(app).post(PENALTY_DETAILS_PAGE_URI)
-                .send(penaltyDetails)
+                .send(penaltyIdentifier)
                 .expect(response => {
                     expect(response.status).to.be.equal(UNPROCESSABLE_ENTITY);
                     expect(response.text).to.contain(pageHeading)
@@ -106,13 +106,13 @@ describe('PenaltyDetailsController', () => {
         });
 
         it('should return 400 when posting empty company number', async () => {
-            const penaltyDetails: PenaltyReferenceDetails = {
+            const penaltyIdentifier: PenaltyIdentifier = {
                 penaltyReference: 'A12345678',
                 companyNumber: ''
             };
 
             await request(app).post(PENALTY_DETAILS_PAGE_URI)
-                .send(penaltyDetails)
+                .send(penaltyIdentifier)
                 .expect(response => {
                     expect(response.status).to.be.equal(UNPROCESSABLE_ENTITY);
                     expect(response.text).to.contain(pageHeading)
@@ -121,13 +121,13 @@ describe('PenaltyDetailsController', () => {
         });
 
         it('should return 400 when posting invalid company number', async () => {
-            const penaltyDetails: PenaltyReferenceDetails = {
+            const penaltyIdentifier: PenaltyIdentifier = {
                 penaltyReference: 'A12345678',
                 companyNumber: 'AB66666666'
             };
 
             await request(app).post(PENALTY_DETAILS_PAGE_URI)
-                .send(penaltyDetails)
+                .send(penaltyIdentifier)
                 .expect(response => {
                     expect(response.status).to.be.equal(UNPROCESSABLE_ENTITY);
                     expect(response.text).to.contain(pageHeading)
